# Fix bug with Get() where non-last key is not an object or array